feat: спринт 14. попытка 1.

+++ src/main/java/ru/practicum/shareit/exceptions/ItemNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.practicum.shareit.exceptions;
 
 public class ItemNotFoundException extends NotFoundException {
-    public ItemNotFoundException(final int itemId) {
+    public ItemNotFoundException(final long itemId) {
         super("Not found item " + itemId);
     }
 }
